Remove legacy setup command module
